src/backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-cosmos
azure-functions
azure.identity
azure-search-documents
azurefunctions.extensions.bindings.blob
langgraph
langchain_community
langchain_openai
python-dotenv